Driver accept bid (#36)
* Add delete for User and Advertisement.

* Expose delUserAcct

* Updated listBidsForRide to return everything in bid.

* Updated carRide.hbs to include select bidder.

* Updated bid.js, ride.js, driver.js.

Now able to select ride. The list of pending rides will be refreshed. Likewise for confirmed rides.

* Updated bid.js, ride.js, rider.js and rider.hbs.

Driver now able to select the bid he wants and it'll become successful. The remaining bids will be converted to unsuccessful. Updated such that Car is displayed as well. However, the car for list available rides are not up yet.

* Updated userId to req.user.

* b.bidStatus = 'pending'.

Changed from b.bidStatus LIKE %pending%